backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Chess Engines and Logic
python-chess==1.999
stockfish==3.28.0

# LLM Integration
ollama==0.1.6
httpx>=0.25.2,<0.26.0

# Lichess API
berserk==0.13.2

# Chess.com API
chess-com-api==1.0.0

# Database (SQLite/PostgreSQL support)
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Configuration
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Utilities
python-dateutil==2.8.2
pytz==2023.3
